Cleans up API models and renames them to a more significative name
