+++ MLMini/Screens/ProductDetailView/Model/ItemDetailInfraestructureModel.swift
@@ -14,15 +14,12 @@ struct ItemDetailInfraestructureModel: Codable {
     let sellerID: Int
     let categoryID: String
     let price: Double
-    let initialQuantity, availableQuantity, soldQuantity: Int
+    let availableQuantity, soldQuantity: Int
     let condition: String
     let permalink: String
     let thumbnail: String
-    let secureThumbnail: String
     let pictures: [Picture]
     let descriptions: [Description]
-    let acceptsMercadopago: Bool
-    let status: String
     
     enum CodingKeys: String, CodingKey {
         case id
@@ -30,35 +27,11 @@ struct ItemDetailInfraestructureModel: Codable {
         case sellerID = "seller_id"
         case categoryID = "category_id"
         case price
-        case initialQuantity = "initial_quantity"
         case availableQuantity = "available_quantity"
         case soldQuantity = "sold_quantity"
         case condition, permalink, thumbnail
-        case secureThumbnail = "secure_thumbnail"
         case pictures
         case descriptions
-        case acceptsMercadopago = "accepts_mercadopago"
-        case status
-        
-    }
-    
-    init(id: String, title: String, sellerID: Int, categoryID: String, price: Double,  initialQuantity: Int, availableQuantity: Int, soldQuantity: Int, condition: String, permalink: String, thumbnail: String, secureThumbnail: String, pictures: [Picture], descriptions: [Description], acceptsMercadopago: Bool, status: String) {
-        self.id = id
-        self.title = title
-        self.sellerID = sellerID
-        self.categoryID = categoryID
-        self.price = price
-        self.initialQuantity = initialQuantity
-        self.availableQuantity = availableQuantity
-        self.soldQuantity = soldQuantity
-        self.condition = condition
-        self.permalink = permalink
-        self.thumbnail = thumbnail
-        self.secureThumbnail = secureThumbnail
-        self.pictures = pictures
-        self.descriptions = descriptions
-        self.acceptsMercadopago = acceptsMercadopago
-        self.status = status
     }
 }
 //MARK: - ToModel
