+++ MLMini/Screens/ProductTableView/Model/ItemRowDomainModel.swift
@@ -1,0 +1,15 @@
+//
+//  ItemRowDomainModel.swift
+//  MLMini
+//
+//  Created by Alexis Garcia Carli on 23/06/2023.
+//  Copyright © 2023 ale. All rights reserved.
+//
+
+struct ItemRowDomainModel {
+    let id: String
+    let price: Double
+    let title: String
+    let thumbnail: String
+    let freeShipping: Bool
+}
