+++ MLMiniTests/Screens/ProductTableView/Mocks/ProductListRepositoryMock.swift
@@ -11,20 +11,20 @@ import Foundation
 
 struct ProductListRepositoryMock {
     
-    static func createResponseFrom(json nameOfFile: String) -> APIResponseModel? {
+    static func createResponseFrom(json nameOfFile: String) -> ProductListInfraestructureModel? {
         guard let path = Bundle(for: MockSwift.self).path(forResource: nameOfFile, ofType: "json") else { return nil }
         do {
             let json = try Data(contentsOf: URL(fileURLWithPath: path), options: .alwaysMapped)
             let decoder = JSONDecoder()
-            guard let response = try? decoder.decode(APIResponseModel.self, from: json) else { return nil}
+            guard let response = try? decoder.decode(ProductListInfraestructureModel.self, from: json) else { return nil}
             return response
         } catch { return nil }
     }
     
-    static func normalApiModel() -> APIResponseModel {
+    static func normalApiModel() -> ProductListInfraestructureModel {
         let paging = Paging(total: 100, offset: 10, limit: 50, primaryResults: 5)
         let results = [
-            Result(id: "01",
+            ProductResult(id: "01",
                    title: "Headset1",
                    price: 43.0,
                    availableQuantity: 5,
@@ -39,12 +39,12 @@ struct ProductListRepositoryMock {
                    catalogProductID: nil)
         ]
         
-        return APIResponseModel(query: "headsets", paging: paging, results: results)
+        return ProductListInfraestructureModel(query: "headsets", paging: paging, results: results)
     }
     
-    static func emptyResultsApiModel() -> APIResponseModel {
+    static func emptyResultsApiModel() -> ProductListInfraestructureModel {
         let paging = Paging(total: 100, offset: 10, limit: 50, primaryResults: 5)
-        let results = [Result]()
-        return APIResponseModel(query: "headsets", paging: paging, results: results)
+        let results = [ProductResult]()
+        return ProductListInfraestructureModel(query: "headsets", paging: paging, results: results)
     }
 }
