+++ MLMiniTests/Screens/ProductTableView/Mocks/ProductListRepositorySpy.swift
@@ -11,16 +11,16 @@ import Foundation
 @testable import MLMini
 
 class ProductListRepositorySpy: ProductListRepositoryProtocol, Mock {
-    var stubbedApiResponseModel: APIResponseModel?
+    var stubbedApiResponseModel: ProductListInfraestructureModel?
     var stubbedErrorType: NetworkError = .invalidUrl
     var didCallGetProductListData = MockCounter()
     var itemName: String?
     
-    func getProductListData(productName: String) -> AnyPublisher<APIResponseModel, NetworkError> {
+    func getProductListData(productName: String) -> AnyPublisher<ProductListInfraestructureModel, NetworkError> {
         didCallGetProductListData.wasCalled()
         itemName = productName
         guard let apiResponseModel = stubbedApiResponseModel else {
-            return AnyPublisher(Fail<APIResponseModel, NetworkError>(error: stubbedErrorType))
+            return AnyPublisher(Fail<ProductListInfraestructureModel, NetworkError>(error: stubbedErrorType))
         }
         return Future { promise in
             promise(.success(apiResponseModel))
